Export SIG config version as uint

+++ go/sig/config/config.go
@@ -29,7 +29,7 @@ import (
 // Cfg is a direct Go representation of the JSON file format.
 type Cfg struct {
 	ASes          map[addr.ISD_AS]*ASEntry
-	ConfigVersion float64
+	ConfigVersion uint64
 }
 
 // Load a JSON config file from path and parse it into a Cfg struct.
